requirements.txt
python-dotenv
langchain
openai
google-api-python-client # Google API
google-search-results # SerpAPI
youtube-transcript-api
tiktoken
faiss-cpu
wikipedia
psycopg2
pinecone-client
pgvector